requirements.txt
pip
streamlit
pandas
plotly_express
Pillow  

